requirements.txt
mysql-connector-python~=8.0.32
opencv-python~=4.7.0.72
numpy~=1.24.2
pandas~=1.5.3
joblib~=1.2.0
Flask~=2.2.3
scikit-learn~=1.2.1